Split StoreDbContext into interface and implementation
