+++ Store.Data/StoreDbContext.cs
@@ -4,34 +4,15 @@ using Store.Data.ModelConfigurations;
 using Store.Data.Extensions;
 using Store.Data.Entities;
 
-using Npgsql;
-
 namespace Store.Data
 {
-    public class StoreDbContext : DbContext
+    public abstract class StoreDbContext : DbContext
     {
         public DbSet<Product> Products { get; set; }
         public DbSet<Manufacturer> Manufacturers { get; set; }
         public DbSet<User> Users { get; set; }
         public DbSet<AccountHistory> AccountHistories { get; set; }
         public DbSet<EventTypeInfo> EventTypeInfo { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            var csb = new NpgsqlConnectionStringBuilder
-            {
-                Database = "store",
-                Host = "localhost",
-                Username = "postgres",
-                Password = "admin"
-            };
-
-            var connectionString = csb.ToString();
-
-            optionsBuilder
-                .UseNpgsql(connectionString)
-                .UseCamelCaseNamingConvention();
-        }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
